refactor: fold code-review cleanups — triple-return resolver, shared test helpers, sorted load_tools serialization
